+++ web-frontend/modules/core/store/auth.js
@@ -81,7 +81,13 @@ export const actions = {
       setToken(data.token, this.app)
       commit('SET_USER_DATA', data)
       dispatch('startRefreshTimeout')
-    } catch {
+    } catch (error) {
+      // If the server can't be reached because of a network error we want to
+      // fail hard so that the correct error message is shown.
+      if (!error.status) {
+        throw error
+      }
+
       // The token could not be refreshed, this means the token is no longer
       // valid and the user not logged in anymore.
       unsetToken(this.app)