+++ changelog.md
@@ -13,6 +13,7 @@
   another field.
 * Fixed bug where the Editable component was not working if a prent a user-select:
   none; property.
+* Fail hard when the web-frontend can't reach the backend because of a network error.
 
 ## Released (2021-02-04)